# Cache common int to string conversions