--- a/Jint/Native/RegExp/RegExpPrototype.cs
+++ b/Jint/Native/RegExp/RegExpPrototype.cs
@@ -71,7 +71,7 @@ namespace Jint.Native.RegExp
             var lastIndex = TypeConverter.ToNumber(R.Get("lastIndex"));
             var i = TypeConverter.ToInteger(lastIndex);
             var global = R.Global;
-            
+
             if (!global)
             {
                 i = 0;
@@ -101,7 +101,7 @@ namespace Jint.Native.RegExp
             }
 
             var e = r.Index + r.Length;
-            
+
             if (global)
             {
                 R.Put("lastIndex", (double) e, true);
@@ -110,12 +110,12 @@ namespace Jint.Native.RegExp
             var matchIndex = r.Index;
 
             var a = InitReturnValueArray(Engine.Array.Construct(Arguments.Empty), s, n, matchIndex);
-            
+
             for (var k = 0; k < n; k++)
             {
                 var group = r.Groups[k];
                 var value = group.Success ? group.Value : Undefined.Instance;
-                a.DefineOwnProperty(k.ToString(), new PropertyDescriptor(value, true, true, true), true);            
+                a.DefineOwnProperty(TypeConverter.ToString(k), new PropertyDescriptor(value, true, true, true), true);
             }
 
             return a;