… Jint/Native/String/StringPrototype.cs