Now a separate lookup that is used for integers. Memory usage drops by third when using latest array performance benchmark.

``` ini
BenchmarkDotNet=v0.10.11, OS=Windows 10 Redstone 3 [1709, Fall Creators Update] (10.0.16299.125)
Processor=Intel Core i7-6820HQ CPU 2.70GHz (Skylake), ProcessorCount=8
Frequency=2648437 Hz, Resolution=377.5812 ns, Timer=TSC
.NET Core SDK=2.1.2
  [Host]     : .NET Core 2.0.3 (Framework 4.6.25815.02), 64bit RyuJIT
  Job-LPPWMI : .NET Core 2.0.3 (Framework 4.6.25815.02), 64bit RyuJIT

InvocationCount=4  LaunchCount=1  TargetCount=3  
UnrollFactor=4  WarmupCount=3  
```
### Before

| Method |  N | ReuseEngine |    Mean |   Error |   StdDev |       Gen 0 |      Gen 1 |      Gen 2 | Allocated |
|------- |--- |------------ |--------:|--------:|---------:|------------:|-----------:|-----------:|----------:|
|   **Jint** | **20** |       **False** | **4.169 s** | **2.139 s** | **0.1209 s** | **784750.0000** | **40500.0000** | **17500.0000** |   **3.18 GB** |
|   **Jint** | **20** |        **True** | **4.090 s** | **1.327 s** | **0.0750 s** | **786750.0000** | **40250.0000** | **18500.0000** |   **3.17 GB** |

### After

| Method |  N | ReuseEngine |    Mean |    Error |   StdDev |       Gen 0 |      Gen 1 |      Gen 2 | Allocated |
|------- |--- |------------ |--------:|---------:|---------:|------------:|-----------:|-----------:|----------:|
|   **Jint** | **20** |       **False** | **2.586 s** | **0.2367 s** | **0.0134 s** | **504000.0000** | **27500.0000** | **13500.0000** |   **2.06 GB** |
|   **Jint** | **20** |        **True** | **2.663 s** | **1.9397 s** | **0.1096 s** | **504000.0000** | **24000.0000** | **11750.0000** |   **2.06 GB** |