@lahma 2017-12-19
Hopefully now the sweet spot for minimal changes for handling the common string/char caching. I added some comments to clarify some decisions. If we'd get this merged I could continue with the next effort.

/cc @ayende @sebastienros