@sebastienros 2017-12-19
LGTM minus comments.