@lahma 2017-12-19
@sebastienros I've done the changes you suggested, thanks for the quick review and good suggestions. I made the string cache a tad smaller now as we are allocation eagerly - probably can grow it later if needed.

Let me know if you want more changes.